+++ src/cmd/commands.hpp
@@ -91,6 +91,8 @@ class Commands {
 
   void addToList(const std::string& seq_id, const std::string& list_filename);
 
+  void exportFormulas(const std::string& output_file);
+
   void commitAddedPrograms(size_t min_commit_count = 5);
 
   void commitUpdatedAndDeletedPrograms();